+++ app/src/main/java/development/iarratais/fragment/SIMFragment.java
@@ -7,19 +7,24 @@
 package development.iarratais.fragment;
 
 
+import android.app.Fragment;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import development.iarratais.androidinfo.R;
+import development.iarratais.utils.SIMInfoUtil;
 
 /**
  * A simple {@link Fragment} subclass.
  */
 public class SIMFragment extends Fragment {
 
+    View rootView;
+
+    SIMInfoUtil simInfoUtil;
 
     public SIMFragment() {
         // Required empty public constructor
@@ -29,8 +34,16 @@ public class SIMFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_sim, container, false);
+
+        rootView = inflater.inflate(R.layout.fragment_sim, container, false);
+        simInfoUtil = new SIMInfoUtil(getActivity());
+
+        Log.d(getClass().getSimpleName(), "SIM Number: " + simInfoUtil.getSIMNumber());
+        Log.d(getClass().getSimpleName(), "Operator: " + simInfoUtil.getNetworkOperator());
+        Log.d(getClass().getSimpleName(), "Operator country: " + simInfoUtil
+                .getNetworkOperatorCountry());
+
+        return rootView;
     }
 
 }